refactor command to new folder
